project done

+++ Backend/app/Actions/Comments/DeleteCommentAction.php
@@ -4,7 +4,9 @@ namespace App\Actions\Comments;
 
 use App\Models\Comment;
 use Illuminate\Database\Eloquent\ModelNotFoundException;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Facades\Log;
+use Illuminate\Validation\UnauthorizedException;
 
 class DeleteCommentAction
 {
@@ -15,6 +17,11 @@ class DeleteCommentAction
         if (!$comment) {
             throw new ModelNotFoundException("Comment not found");
             Log::info('Comment not found');
+        }
+        if (!Gate::allows('delete', $comment)) {
+            throw new UnauthorizedException(
+                message: "You don't have permission to update this Comment"
+            );
         }
         try {
             $comment->delete();
